Move User class to application target
